+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-10-02
+### Added
+- New variables, `consul_dc` and `nomad_dc` for the Consul and Nomd datacenter names respectively
+### Removed
+- `environment` variable
+
 ## [0.2.0] - 2023-09-30
 ### Added
 - Custom service account for Compute Engine